Allow entries to be updated without having to rebuild entire database
Also added a silent mode which is enabled by using the options -s or --silent; any changes to the database will not be displayed.

Some validation of the Upload Date and MEMORY REQUIRED fields in the CSV file are now performed to prevent errors from occurring when invalid entries are uploaded to the database.

Some code was also refactored and/or moved to its own function.